design : DateNotification 컴포넌트 기능 구현
