+++ package.json
@@ -30,6 +30,7 @@
     "eslint-plugin-react-refresh": "^0.4.5",
     "prettier": "^3.1.1",
     "react-router-dom": "^6.21.1",
+    "sass": "^1.69.7",
     "styled-components": "^6.1.6",
     "vite": "^5.0.8"
   }
