+++ src/components/CalendarPage/DateNotification.css
@@ -1,0 +1,4 @@
+.swal2-toast {
+  font-size: 18px;
+  max-width: 500px;
+}
